feat: removed oldSearchSavingAutosuggest
